+++ cli/tests/test_describe_command.rs
@@ -12,6 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use std::path::Path;
+
+use indoc::indoc;
+
 use crate::common::{get_stderr_string, TestEnvironment};
 
 #[test]
@@ -172,6 +176,180 @@ fn test_describe() {
 }
 
 #[test]
+fn test_describe_multiple_commits() {
+    let mut test_env = TestEnvironment::default();
+    test_env.jj_cmd_ok(test_env.env_root(), &["git", "init", "repo"]);
+    let repo_path = test_env.env_root().join("repo");
+
+    let edit_script = test_env.set_up_fake_editor();
+
+    // Initial setup
+    test_env.jj_cmd_ok(&repo_path, &["new"]);
+    test_env.jj_cmd_ok(&repo_path, &["new"]);
+    insta::assert_snapshot!(get_log_output(&test_env, &repo_path), @r###"
+    @  c6349e79bbfd
+    ◉  65b6b74e0897
+    ◉  230dd059e1b0
+    ◉  000000000000
+    "###);
+
+    // Set the description of multiple commits using `-m` flag
+    let (stdout, stderr) = test_env.jj_cmd_ok(
+        &repo_path,
+        &["describe", "@", "@--", "-m", "description from CLI"],
+    );
+    insta::assert_snapshot!(stdout, @"");
+    insta::assert_snapshot!(stderr, @r###"
+    Updated 2 commits
+    Rebased 1 descendant commits
+    Working copy now at: kkmpptxz 3b1c88bc (empty) description from CLI
+    Parent commit      : rlvkpnrz 4026cb14 (empty) (no description set)
+    "###);
+    insta::assert_snapshot!(get_log_output(&test_env, &repo_path), @r###"
+    @  3b1c88bccc80 description from CLI
+    ◉  4026cb14df00
+    ◉  65507340cfda description from CLI
+    ◉  000000000000
+    "###);
+
+    // Check that the text file gets initialized with the current description of
+    // each commit and doesn't update commits if no changes are made.
+    // Commit descriptions are edited in topological order
+    std::fs::write(&edit_script, "dump editor0").unwrap();
+    let (stdout, stderr) = test_env.jj_cmd_ok(&repo_path, &["describe", "@", "@-"]);
+    insta::assert_snapshot!(stdout, @"");
+    insta::assert_snapshot!(stderr, @r###"
+    Nothing changed.
+    "###);
+    insta::assert_snapshot!(
+        std::fs::read_to_string(test_env.env_root().join("editor0")).unwrap(), @r###"
+    JJ: describe 4026cb14df00
+
+    JJ: describe 3b1c88bccc80
+    description from CLI
+
+    JJ: Lines starting with "JJ: " (like this one) will be removed.
+    "###);
+
+    // Set the description of multiple commits in the editor
+    std::fs::write(
+        &edit_script,
+        indoc! {"
+            write
+            JJ: describe 4026cb14df00
+            description from editor of @-
+
+            further commit message of @-
+
+            JJ: describe 3b1c88bccc80
+            description from editor of @
+
+            further commit message of @
+
+            JJ: Lines starting with \"JJ: \" (like this one) will be removed.
+        "},
+    )
+    .unwrap();
+    let (stdout, stderr) = test_env.jj_cmd_ok(&repo_path, &["describe", "@", "@-"]);
+    insta::assert_snapshot!(stdout, @"");
+    insta::assert_snapshot!(stderr, @r###"
+    Updated 2 commits
+    Working copy now at: kkmpptxz 83aad183 (empty) description from editor of @
+    Parent commit      : rlvkpnrz 1e274f89 (empty) description from editor of @-
+    "###);
+    insta::assert_snapshot!(get_log_output(&test_env, &repo_path), @r###"
+    @  83aad18359bf description from editor of @
+    │
+    │  further commit message of @
+    ◉  1e274f899cb2 description from editor of @-
+    │
+    │  further commit message of @-
+    ◉  65507340cfda description from CLI
+    ◉  000000000000
+    "###);
+
+    // Fails if the edited message has a commit has multiple descriptions
+    std::fs::write(
+        &edit_script,
+        indoc! {"
+            write
+            JJ: describe 1e274f899cb2
+            first description from editor of @-
+
+            further commit message of @-
+
+            JJ: describe 1e274f899cb2
+            second description from editor of @-
+
+            further commit message of @-
+
+            JJ: describe 83aad18359bf
+            updated description from editor of @
+
+            further commit message of @
+
+            JJ: Lines starting with \"JJ: \" (like this one) will be removed.
+        "},
+    )
+    .unwrap();
+    let stderr = test_env.jj_cmd_failure(&repo_path, &["describe", "@", "@-"]);
+    insta::assert_snapshot!(stderr, @r###"
+    Error: The following commits were found in the edited message multiple times: 1e274f899cb2
+    "###);
+
+    // Fails if the edited message has unexpected commit IDs
+    std::fs::write(
+        &edit_script,
+        indoc! {"
+            write
+            JJ: describe 000000000000
+            unexpected commit ID
+
+            JJ: describe 1e274f899cb2
+            description from editor of @-
+
+            further commit message of @-
+
+            JJ: describe 83aad18359bf
+            description from editor of @
+
+            further commit message of @
+
+            JJ: Lines starting with \"JJ: \" (like this one) will be removed.
+        "},
+    )
+    .unwrap();
+    let stderr = test_env.jj_cmd_failure(&repo_path, &["describe", "@", "@-"]);
+    insta::assert_snapshot!(stderr, @r###"
+    Error: The following commits were not being edited, but were found in the edited message: 000000000000
+    "###);
+
+    // Fails if the edited message has missing commit messages
+    std::fs::write(
+        &edit_script,
+        indoc! {"
+            write
+            JJ: describe 83aad18359bf
+            description from editor of @
+
+            further commit message of @
+
+            JJ: Lines starting with \"JJ: \" (like this one) will be removed.
+        "},
+    )
+    .unwrap();
+    let stderr = test_env.jj_cmd_failure(&repo_path, &["describe", "@", "@-"]);
+    insta::assert_snapshot!(stderr, @r###"
+    Error: The description for the following commits were not found in the edited message: 1e274f899cb2
+    "###);
+
+    // Fails if the editor fails
+    std::fs::write(&edit_script, "fail").unwrap();
+    let stderr = test_env.jj_cmd_failure(&repo_path, &["describe", "@", "@-"]);
+    assert!(stderr.contains("exited with an error"));
+}
+
+#[test]
 fn test_multiple_message_args() {
     let test_env = TestEnvironment::default();
     test_env.jj_cmd_ok(test_env.env_root(), &["git", "init", "repo"]);
@@ -293,19 +471,30 @@ fn test_describe_author() {
             &repo_path,
             &[
                 "log",
-                "-r@",
+                "-r..",
                 "-T",
                 r#"format_signature(author) ++ "\n" ++ format_signature(committer)"#,
             ],
         )
     };
+
+    // Initial setup
+    test_env.jj_cmd_ok(&repo_path, &["new"]);
+    test_env.jj_cmd_ok(&repo_path, &["new"]);
+    test_env.jj_cmd_ok(&repo_path, &["new"]);
     insta::assert_snapshot!(get_signatures(), @r###"
-    @  Test User test.user@example.com 2001-02-03 04:05:07.000 +07:00
+    @  Test User test.user@example.com 2001-02-03 04:05:10.000 +07:00
+    │  Test User test.user@example.com 2001-02-03 04:05:10.000 +07:00
+    ◉  Test User test.user@example.com 2001-02-03 04:05:09.000 +07:00
+    │  Test User test.user@example.com 2001-02-03 04:05:09.000 +07:00
+    ◉  Test User test.user@example.com 2001-02-03 04:05:08.000 +07:00
+    │  Test User test.user@example.com 2001-02-03 04:05:08.000 +07:00
+    ◉  Test User test.user@example.com 2001-02-03 04:05:07.000 +07:00
     │  Test User test.user@example.com 2001-02-03 04:05:07.000 +07:00
     ~
     "###);
 
-    // Reset the author (the committer is always reset)
+    // Reset the author for the latest commit (the committer is always reset)
     test_env.jj_cmd_ok(
         &repo_path,
         &[
@@ -318,8 +507,45 @@ fn test_describe_author() {
         ],
     );
     insta::assert_snapshot!(get_signatures(), @r###"
-    @  Ove Ridder ove.ridder@example.com 2001-02-03 04:05:09.000 +07:00
-    │  Ove Ridder ove.ridder@example.com 2001-02-03 04:05:09.000 +07:00
+    @  Ove Ridder ove.ridder@example.com 2001-02-03 04:05:12.000 +07:00
+    │  Ove Ridder ove.ridder@example.com 2001-02-03 04:05:12.000 +07:00
+    ◉  Test User test.user@example.com 2001-02-03 04:05:09.000 +07:00
+    │  Test User test.user@example.com 2001-02-03 04:05:09.000 +07:00
+    ◉  Test User test.user@example.com 2001-02-03 04:05:08.000 +07:00
+    │  Test User test.user@example.com 2001-02-03 04:05:08.000 +07:00
+    ◉  Test User test.user@example.com 2001-02-03 04:05:07.000 +07:00
+    │  Test User test.user@example.com 2001-02-03 04:05:07.000 +07:00
     ~
     "###);
+
+    // Reset the author for multiple commits (the committer is always reset)
+    test_env.jj_cmd_ok(
+        &repo_path,
+        &[
+            "describe",
+            "@---",
+            "@-",
+            "--config-toml",
+            r#"user.name = "Ove Ridder"
+            user.email = "ove.ridder@example.com""#,
+            "--no-edit",
+            "--reset-author",
+        ],
+    );
+    insta::assert_snapshot!(get_signatures(), @r###"
+    @  Ove Ridder ove.ridder@example.com 2001-02-03 04:05:12.000 +07:00
+    │  Ove Ridder ove.ridder@example.com 2001-02-03 04:05:14.000 +07:00
+    ◉  Ove Ridder ove.ridder@example.com 2001-02-03 04:05:14.000 +07:00
+    │  Ove Ridder ove.ridder@example.com 2001-02-03 04:05:14.000 +07:00
+    ◉  Test User test.user@example.com 2001-02-03 04:05:08.000 +07:00
+    │  Ove Ridder ove.ridder@example.com 2001-02-03 04:05:14.000 +07:00
+    ◉  Ove Ridder ove.ridder@example.com 2001-02-03 04:05:14.000 +07:00
+    │  Ove Ridder ove.ridder@example.com 2001-02-03 04:05:14.000 +07:00
+    ~
+    "###);
+}
+
+fn get_log_output(test_env: &TestEnvironment, repo_path: &Path) -> String {
+    let template = r#"commit_id.short() ++ " " ++ description"#;
+    test_env.jj_cmd_success(repo_path, &["log", "-T", template])
 }